Add MWE for allocating and accessing device-side memory
Question regarding device-side memory outside of SYCL kernels:
https://software.intel.com/en-us/forums/oneapi-data-parallel-c-compiler/topic/856418

When running on a machine with an Intel iGPU, memory allocated with
malloc_device is accessible outside a kernel -- this is unexpected.
Turns out, this *is* expected since the Intel Unified Memory Architecture
uses shared memory for the CPU and iGPU; see:
https://software.intel.com/content/dam/develop/public/us/en/documents/the-architecture-of-intel-processor-graphics-gen11-r1new.pdf

When compiling this code with Intel's llvm for CUDA, a segfault occurs when
attempting to access the malloc_device memory outside a kernel (as expected).
This is checked by instead using a cl::sycl::stream to output the data in
device-side memory.